Code review changes

+++ Konteh/Konteh.Infrastructure/Repositories/IQuestionRepository.cs
@@ -4,6 +4,6 @@ namespace Konteh.Infrastructure.Repositories
 {
     public interface IQuestionRepository : IRepository<Question>
     {
-        Task<(IEnumerable<Question> SearchedQuestions, int QuestionCount)> PaginateItems(int page, float pageSize, string? questionText = null);
+        Task<(IEnumerable<Question> SearchedQuestions, int QuestionCount)> PaginateItems(int page, int pageSize, string? questionText = null);
     }
 }
